правки перед итоговой проверкой
