+++ src/index.html
@@ -53,12 +53,16 @@
             </section>
             <section class="about-author">
                 <div class="about-author__photo-wrapper">
-                    <img class="about-author__photo" src="<%=require('./images/image-03.png')%>" alt="мое фото">
+                    <img class="about-author__photo" src="<%=require('./images/my-photo.png')%>" alt="мое фото">
                 </div>
                 <div class="about-author__text">
                     <h2 class="about-author__title">Об авторе</h2>
-                    <p class="about-author__paragraph">Это блок с описаением автора проекта. Здесь следует указать, как вас зовут, чем вы занимаетесь, какими технологиями разработки владете. </p>
-                    <p class="about-author__paragraph">Также можно рассказать о процессе обучения в Практикуме, чему вы тут научились, и чем можете помочь потенциальным заказчикам. Если хотите, оставьте свои контакты: email и телефон. Так заказчики смогут с вами связаться.</p>
+                    <p class="about-author__paragraph">Меня зовут Кудрявцева Ксения и я начинающий frontend разработчик. В ходе обучения на Яндекс.Практикуме мною были освоены технологии HTML5, CSS3 и JavaScript ES6. </p>
+                    <p class="about-author__paragraph">Имею опыт работы с семантическими тегами HTML и адаптивной (flex-box, grid), при этом валидной, версткой. При необходимости могу адаптировать проекты для БЭМ структуры: как для CSS, так и для JS.</p>
+                    <p class="about-author__paragraph">Мой опыт в JS не ограничивается простым взаимодействием с DOM элементами страницы, умею работать с API и асинхронным ООП кодом.</p>
+                    <p class="about-author__paragraph">В плане верстки есть опыт pixel-perfect разработки более десятка проектов сделанных в Figma.</p>
+                    <p class="about-author__paragraph">Современная frontend разработка подразумевает еще и работу с npm пакетами и Webpack'ом в частности. Кстати, данный проект был собран, а код транспилирован и минифицирован с его помощью.</p>
+                    <p class="about-author__paragraph">Когда обычная документация не дает искомых ответов, не стесняюсь пользоваться гуглом(дальше 1 страницы).</p>
                 </div>
             </section>
         </main>
