+++ src/js/modules/variebles.js
@@ -7,3 +7,7 @@ export const retrievedNewsArray = localStorage.getItem('newsArray');
 export const retrievedQuery = localStorage.getItem('query');
 export const monthNamesRussian = ["Января", "Февраля", "Марта", "Апреля", "Мая", "Июня", "Июля", "Августа", "Сентября", "Октября", "Ноября", "Декабря"];
 export const regexDateTransform = /(\d{1,4})-(\d{1,2})-(\d{1,2})/g;
+
+export const today = new Date(Date.now() + 60 * 60 * 1000).toLocaleString().match(regexOne).join('').replace(regexTwo,(match, date, month, year) => `${year}-${month}-${date}`);
+//Получение даты 7 дней назад
+export const pastWeek = new Date(Date.now() - 6 * 24 * 60 * 60 * 1000).toLocaleString().match(regexOne).join('').replace(regexTwo,(match,date,month,year) => `${year}-${month}-${date}`);
